refactor(httpapi): split groups from handlers
